Use correct file format when creating Hive tables.
Previously, when the prepare_benchmark.py script created Hive tables, it
hardcoded a specific file format instead of using the file format
specified by the "--file-format" command line parameter. This led to
parsing problems because the actual table data that was copied from S3
was in the format specified by "--file-format", which could potentially
be different than the hardcoded format. This commit fixes this problem
by changing all of the Hive table creation commands to use the
user-specified file format.