Use the correct delimiter for compressed sequence files
Previously, the prepare_benchmark.py script hardcoded the column
delimiter to use. Unfortunately, this was the wrong delimiter for
compressed sequence files. This commit removes the hardcoded column
delimiter for compressed sequence files, causing the default delimiter
to be used. This fix has only been applied to the Spark preparation
code.